Move initial state to another module
