+++ client/state/domains/nameservers/initial.js
@@ -1,0 +1,6 @@
+export default {
+	isFetching: false,
+	hasLoadedFromServer: false,
+	error: false,
+	list: null,
+};
