+++ client/state/domains/nameservers/selectors.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { initialDomainState } from './reducer';
+import initialDomainState from './initial';
 
 export function getNameserversByDomainName( state, domainName ) {
 	return state.domains?.nameservers[ domainName ] || initialDomainState;
