+++ client/state/domains/nameservers/test/selectors.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { initialDomainState } from '../reducer';
+import initialDomainState from '../initial';
 import { getNameserversByDomainName } from '../selectors';
 
 describe( 'selectors', () => {
